+++ src/chevah_compat/tests/elevated/test_filesystem.py
@@ -614,8 +614,8 @@ class TestSymbolicLinks(OSAccountFileSystemTestCase, SymbolicLinksMixin):
         rights = ()
 
     CREATE_TEST_USER = TestUser(
-        name=mk.string(),
-        password=mk.string(),
+        name=mk.ascii(),
+        password=mk.password(),
         home_group=TEST_ACCOUNT_GROUP,
         posix_uid=mk.posixID(),
         windows_required_rights=rights,